actions/requirements-actions.txt
requests~=2.25.1
rasa-sdk~=2.8.0
elasticsearch-dsl>=7.0.0,<8.0.0
python-dotenv~=0.19.0
pytest~=6.2.4
pytest-asyncio~=0.15.1
thefuzz~=0.19.0
